# dennic_admin_api_gateway/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Dennic Admin Project",
        "contact": {},
        "version": "1.7"
    },
    "host": "admin.dennic.uz",
    "paths": {
...
        "/v1/doctor-working-hours": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "ListDoctorWorkingHours - Api for list doctor_working_hours",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor Working Hours"
                ],
                "summary": "ListDoctorWorkingHours",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "day_of_week"
                        ],
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.ListDoctorWorkingHours"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "UpdateDoctorWorkingHours - Api for update doctor_working_hours",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor Working Hours"
                ],
                "summary": "UpdateDoctorWorkingHours",
                "parameters": [
                    {
                        "description": "UpdateDoctorWorkingHoursReq",
                        "name": "UpdateDoctorWorkingHoursReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorWorkingHoursReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorWorkingHoursRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "DeleteDoctorWorkingHours - Api for delete doctor_working_hours",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor Working Hours"
                ],
                "summary": "DeleteDoctorWorkingHours",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "hard_delete",
                        "name": "hard_delete",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StatusRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/doctor-working-hours/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "CreateDoctorWorkingHours - Api for crete doctor_working_hours",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor Working Hours"
                ],
                "summary": "CreateDoctorWorkingHours",
                "parameters": [
                    {
                        "description": "DoctorServiceReq",
                        "name": "DoctorWorkingHoursReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorWorkingHoursReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorWorkingHoursRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/doctor-working-hours/get": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "GetDoctorWorkingHours - Api for get doctor_working_hours",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor Working Hours"
                ],
                "summary": "GetDoctorWorkingHours",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "is_active",
                        "name": "is_active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorWorkingHoursRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model_common.ResponseError": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model_common.StandardErrorModel": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/model_common.ResponseError"
                }
            }
        },
...
        "model_healthcare_service.DoctorWorkingHoursReq": {
            "type": "object",
            "properties": {
                "day_of_week": {
                    "type": "string",
                    "example": "Monday"
                },
                "doctor_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614274001"
                },
                "finish_time": {
                    "type": "string",
                    "example": "12:00:00"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614274001"
                },
                "start_time": {
                    "type": "string",
                    "example": "12:00:00"
                }
            }
        },
        "model_healthcare_service.DoctorWorkingHoursRes": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "day_of_week": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "finish_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model_healthcare_service.ListDoctorWorkingHours": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "doctor_working_hours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model_healthcare_service.DoctorWorkingHoursRes"
                    }
                }
            }
        },
...
        "models.StatusRes": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}